Files/09/requirements.txt
azure-ai-projects~=2.0.0b1
azure-identity~=1.20.0
openai~=2.0.1
python-dotenv~=1.0.0
azure-monitor-opentelemetry~=1.6.0
opentelemetry-instrumentation-openai-v2>=2.0b0